Changing wait time data type to float

+++ src/unstract/sdk/adapters/embedding/no_op_embedding/src/no_op_embedding.py
@@ -7,6 +7,6 @@ class NoOpEmbedding(MockEmbedding):
 
     embed_dim: int
 
-    def __init__(self, embed_dim: int, wait_time: int, **kwargs: Any) -> None:
+    def __init__(self, embed_dim: int, wait_time: float, **kwargs: Any) -> None:
         """Init params."""
         super().__init__(embed_dim=embed_dim, **kwargs, wait_time=wait_time)
